pyproject.toml
[tool.poetry]
name = "python_boilerplate"
version = "0.1.0"
description = ""
authors = ["Gabriel Fonseca <fonsecacrz7@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.1"
SQLAlchemy = "^1.4.21"
python-dotenv = "^0.18.0"
Cerberus = "^1.3.4"
alembic = "^1.6.5"
PyJWT = "^2.1.0"
passlib = "^1.7.4"
emails = "^0.6"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
flake8 = "^3.9.2"
black = "^21.6b0"
isort = "^5.9.2"
coverage = "^5.5"
pytest = "^6.2.4"
Faker = "^8.11.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
